requirement.txt
asgiref==3.3.4
Django==3.2.3
djangorestframework==3.12.4
mysql==0.0.2
mysqlclient==2.0.3
pytz==2021.1
sqlparse==0.4.1
